requirements.txt
autopep8==1.6.0
click==8.1.3
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==1.4.2
python-dotenv==0.20.0
SQLAlchemy==1.4.36
toml==0.10.2
uuid==1.30
Werkzeug==2.1.2
zipp==3.8.0
